Move freebox base entity to separate module
