+++ homeassistant/components/freebox/entity.py
